+++ pylops_mpi/utils/__init__.py
@@ -1,4 +1,5 @@
 # isort: skip_file
 
+from .benchmark import *
 from .dottest import *
 from .deps import *