+++ tests/test_solver.py
@@ -26,7 +26,6 @@ from pylops_mpi import (
     StackedDistributedArray
 )
 
-np.random.seed(42)
 size = MPI.COMM_WORLD.Get_size()
 rank = MPI.COMM_WORLD.Get_rank()
 
@@ -94,6 +93,8 @@ par4j = {
 )
 def test_cg(par):
     """CG with MPIBlockDiag"""
+    np.random.seed(42)
+
     A = np.ones((par["ny"], par["nx"])) + par[
         "imag"] * np.ones((par["ny"], par["nx"]))
     Aop = MatrixMult(np.conj(A.T) @ A, dtype=par['dtype'])
@@ -139,6 +140,8 @@ def test_cg(par):
 )
 def test_cgls(par):
     """CGLS with MPIBlockDiag"""
+    np.random.seed(42)
+
     A = np.ones((par["ny"], par["nx"])) + par[
         "imag"] * np.ones((par["ny"], par["nx"]))
     Aop = MatrixMult(np.conj(A.T) @ A + 1e-5 * np.eye(par["nx"], dtype=par['dtype']),
@@ -186,6 +189,8 @@ def test_cgls(par):
 )
 def test_cgls_broadcastdata(par):
     """CGLS with broadcasted data vector"""
+    np.random.seed(42)
+
     A = (rank + 1) * np.ones((par["ny"], par["nx"])) + (rank + 2) * par[
         "imag"
     ] * np.ones((par["ny"], par["nx"]))
@@ -232,6 +237,8 @@ def test_cgls_broadcastdata(par):
 )
 def test_cgls_broadcastmodel(par):
     """CGLS with broadcasted model vector"""
+    np.random.seed(42)
+
     A = np.ones((par["ny"], par["nx"])) + par[
         "imag"] * np.ones((par["ny"], par["nx"]))
     Aop = MatrixMult(np.conj(A.T) @ A + 1e-5 * np.eye(par["nx"], dtype=par['dtype']),
@@ -281,6 +288,8 @@ def test_cgls_broadcastmodel(par):
 )
 def test_cg_stacked(par):
     """CG with MPIStackedBlockDiag"""
+    np.random.seed(42)
+
     A = np.ones((par["ny"], par["nx"])) + par[
         "imag"] * np.ones((par["ny"], par["nx"]))
     Aop = MatrixMult(np.conj(A.T) @ A + 1e-5 * np.eye(par["nx"], dtype=par['dtype']),
@@ -344,6 +353,8 @@ def test_cg_stacked(par):
 )
 def test_cgls_stacked(par):
     """CGLS with MPIStackedBlockDiag"""
+    np.random.seed(42)
+
     A = np.ones((par["ny"], par["nx"])) + par[
         "imag"] * np.ones((par["ny"], par["nx"]))
     Aop = MatrixMult(np.conj(A.T) @ A + 1e-5 * np.eye(par["nx"], dtype=par['dtype']),