+++ tutorials_nccl/lsm_nccl.py
@@ -1,55 +1,16 @@
 r"""
-Least-squares Migration
-=======================
-Seismic migration involves the manipulation of seismic data to create an image of subsurface reflectivity.
-
-We solve the inverse of the demigration operator to obtain accurate and detailed sub surface images.
-Regardless of the choice of the modelling operator (i.e., ray-based or full wavefield-based), the
-demigration/migration process can be expressed as a linear operator of such kind:
-
-.. math::
-        d(\mathbf{x_r}, \mathbf{x_s}, t) =
-        w(t) * \int\limits_V G(\mathbf{x}, \mathbf{x_s}, t)
-        G(\mathbf{x_r}, \mathbf{x}, t) m(\mathbf{x})\,\mathrm{d}\mathbf{x}
-
-where :math:`m(\mathbf{x})` is the reflectivity at every location in the subsurface,
-:math:`G(\mathbf{x}, \mathbf{x_s}, t)` and :math:`G(\mathbf{x_r}, \mathbf{x}, t)`
-are the Green's functions from source-to-subsurface-to-receiver and finally  :math:`w(t)`
-is the wavelet. Ultimately, while the Green's functions can be computed in many ways,
-solving this system of equations for the reflectivity model is what we generally
-refer to as Least-squares migration (LSM).
-
-We can easily set up this problem where sources are distributed across different ranks,
-and each :class:`pylops.waveeqprocessing.LSM` is responsible for performing modelling
-with the reflectivity at each rank in the subsurface. In a compact matrix-vector notation,
-this problem can be written as:
-
-.. math::
-    \begin{bmatrix}
-        \mathbf{d}_{1}    \\
-        \mathbf{d}_{2}    \\
-        \vdots    \\
-        \mathbf{d}_{N}
-    \end{bmatrix} =
-    \begin{bmatrix}
-        \mathbf{L}_1    \\
-        \mathbf{L}_2    \\
-        \vdots    \\
-        \mathbf{L}_n
-    \end{bmatrix} m
-
-where :math:`\mathbf{L}_i` is a least-squares modelling operator, :math:`\mathbf{d}_i`
-is the data, and :math:`m` is the broadcasted reflectivity at every location on the subsurface.
-
-In this tutorial, we will use the :class:`pylops_mpi.basicoperators.MPIVStack` to perform
-vertical stacking of LSMs and solve our problem.
-
+Least-squares Migration with NCCL
+=================================
+This tutorial is an extension of the :ref:`sphx_glr_tutorials_lsm.py`
+tutorial where PyLops-MPI is run in multi-GPU setting with GPUs communicating
+via NCCL.
 """
 
 import warnings
 warnings.filterwarnings('ignore')
 
 import numpy as np
+import cupy as cp
 from matplotlib import pyplot as plt
 from mpi4py import MPI
 
@@ -58,14 +19,21 @@ from pylops.waveeqprocessing.lsm import LSM
 
 import pylops_mpi
 
+###############################################################################
+# NCCL communication can be easily initialized with
+# :py:func:`pylops_mpi.utils._nccl.initialize_nccl_comm` operator.
+# One can think of this as GPU-counterpart of :code:`MPI.COMM_WORLD`
+
 np.random.seed(42)
 plt.close("all")
+nccl_comm = pylops_mpi.utils._nccl.initialize_nccl_comm()
 rank = MPI.COMM_WORLD.Get_rank()
-size = MPI.COMM_WORLD.Get_size()
 
 ###############################################################################
-# Let's start with a simple model with two interfaces, where sources are distributed
-# across different ranks.
+# Let's start by defining all the parameters required by the
+# :py:class:`pylops.waveeqprocessing.LSM` operator.
+# Note that this section is exactly the same as the one in the MPI example
+# as we will keep using MPI for transfering metadata (i.e., shapes, dims, etc.)
 
 # Velocity Model
 nx, nz = 81, 60
@@ -125,6 +93,11 @@ if rank == 0:
 # We create a :py:class:`pylops.waveeqprocessing.LSM` at each rank and then push them
 # into a :py:class:`pylops_mpi.basicoperators.MPIVStack` to perform a matrix-vector
 # product with the broadcasted reflectivity at every location on the subsurface.
+# Note that we must use :code:`engine="cuda"` and move the wavelet wav to the GPU prior to creating the operator.
+# Moreover, we allocate the traveltime tables (:code:`lsm.Demop.trav_srcs`, and :code:`lsm.Demop.trav_recs`)
+# to the GPU prior to applying the operator to avoid incurring in the penalty of performing
+# host-to-device memory copies every time the operator is applied. Moreover, we must pass :code:`nccl_comm`
+# to the DistributedArray constructor used to create :code:`refl_dist` in order to use NCCL for communications.
 
 # Wavelet
 nt = 651
@@ -139,15 +112,21 @@ lsm = LSM(
     sources,
     recs,
     v0,
-    wav,
+    cp.asarray(wav.astype(np.float32)),
     wavc,
     mode="analytic",
-    engine="numba",
+    engine="cuda",
+    dtype=np.float32
 )
+lsm.Demop.trav_srcs = cp.asarray(lsm.Demop.trav_srcs.astype(np.float32))
+lsm.Demop.trav_recs = cp.asarray(lsm.Demop.trav_recs.astype(np.float32))
 
 VStack = pylops_mpi.MPIVStack(ops=[lsm.Demop, ])
-refl_dist = pylops_mpi.DistributedArray(global_shape=nx * nz, partition=pylops_mpi.Partition.BROADCAST)
-refl_dist[:] = refl.flatten()
+refl_dist = pylops_mpi.DistributedArray(global_shape=nx * nz,
+                                        partition=pylops_mpi.Partition.BROADCAST,
+                                        base_comm_nccl=nccl_comm,
+                                        engine="cupy")
+refl_dist[:] = cp.asarray(refl.flatten())
 d_dist = VStack @ refl_dist
 d = d_dist.asarray().reshape((nstot, nr, nt))
 
@@ -161,54 +140,62 @@ d_adj = d_adj_dist.asarray().reshape((nstot, nr, nt))
 
 ###############################################################################
 # We calculate the inverse using the :py:func:`pylops_mpi.optimization.basic.cgls`
-# solver.
+# solver. Here, we pass the :code:`nccl_comm` to :code:`x0` to use NCCL as a communicator.
+# In this particular case, the local computation will be done in GPU.
+# Collective communication calls will be carried through NCCL GPU-to-GPU.
 
 # Inverse
 # Initializing x0 to zeroes
-x0 = pylops_mpi.DistributedArray(VStack.shape[1], partition=pylops_mpi.Partition.BROADCAST)
+x0 = pylops_mpi.DistributedArray(VStack.shape[1],
+                                 partition=pylops_mpi.Partition.BROADCAST,
+                                 base_comm_nccl=nccl_comm,
+                                 engine="cupy")
 x0[:] = 0
 minv_dist = pylops_mpi.cgls(VStack, d_dist, x0=x0, niter=100, show=True)[0]
 minv = minv_dist.asarray().reshape((nx, nz))
 d_inv_dist = VStack @ minv_dist
 d_inv = d_inv_dist.asarray().reshape(nstot, nr, nt)
 
-###############################################################################
+##############################################################################
+# Finally we visualize the results. Note that the array must be copied back
+# to the CPU by calling the :code:`get()` method on the CuPy arrays.
+
 if rank == 0:
     # Visualize
     fig1, axs = plt.subplots(1, 3, figsize=(10, 3))
     axs[0].imshow(refl.T, cmap="gray", vmin=-1, vmax=1)
     axs[0].axis("tight")
     axs[0].set_title(r"$m$")
-    axs[1].imshow(madj.T, cmap="gray", vmin=-madj.max(), vmax=madj.max())
+    axs[1].imshow(madj.T.get(), cmap="gray", vmin=-madj.max(), vmax=madj.max())
     axs[1].set_title(r"$m_{adj}$")
     axs[1].axis("tight")
-    axs[2].imshow(minv.T, cmap="gray", vmin=-1, vmax=1)
+    axs[2].imshow(minv.T.get(), cmap="gray", vmin=-1, vmax=1)
     axs[2].axis("tight")
     axs[2].set_title(r"$m_{inv}$")
     plt.tight_layout()
+    fig1.savefig("model.png")
 
     fig2, axs = plt.subplots(1, 3, figsize=(10, 3))
-    axs[0].imshow(d[0, :, :300].T, cmap="gray", vmin=-d.max(), vmax=d.max())
+    axs[0].imshow(d[0, :, :300].T.get(), cmap="gray", vmin=-d.max(), vmax=d.max())
     axs[0].set_title(r"$d$")
     axs[0].axis("tight")
-    axs[1].imshow(d_adj[0, :, :300].T, cmap="gray", vmin=-d_adj.max(), vmax=d_adj.max())
+    axs[1].imshow(d_adj[0, :, :300].T.get(), cmap="gray", vmin=-d_adj.max(), vmax=d_adj.max())
     axs[1].set_title(r"$d_{adj}$")
     axs[1].axis("tight")
-    axs[2].imshow(d_inv[0, :, :300].T, cmap="gray", vmin=-d.max(), vmax=d.max())
+    axs[2].imshow(d_inv[0, :, :300].T.get(), cmap="gray", vmin=-d.max(), vmax=d.max())
     axs[2].set_title(r"$d_{inv}$")
     axs[2].axis("tight")
+    fig2.savefig("data1.png")
 
     fig3, axs = plt.subplots(1, 3, figsize=(10, 3))
-    axs[0].imshow(d[nstot // 2, :, :300].T, cmap="gray", vmin=-d.max(), vmax=d.max())
+    axs[0].imshow(d[nstot // 2, :, :300].T.get(), cmap="gray", vmin=-d.max(), vmax=d.max())
     axs[0].set_title(r"$d$")
     axs[0].axis("tight")
-    axs[1].imshow(d_adj[nstot // 2, :, :300].T, cmap="gray", vmin=-d_adj.max(), vmax=d_adj.max())
+    axs[1].imshow(d_adj[nstot // 2, :, :300].T.get(), cmap="gray", vmin=-d_adj.max(), vmax=d_adj.max())
     axs[1].set_title(r"$d_{adj}$")
     axs[1].axis("tight")
-    axs[2].imshow(d_inv[nstot // 2, :, :300].T, cmap="gray", vmin=-d.max(), vmax=d.max())
+    axs[2].imshow(d_inv[nstot // 2, :, :300].T.get(), cmap="gray", vmin=-d.max(), vmax=d.max())
     axs[2].set_title(r"$d_{inv}$")
     axs[2].axis("tight")
     plt.tight_layout()
-
-###############################################################################
-# To run this tutorial with our NCCL backend, refer to `Least-squares Migration with NCCL tutorial <https://github.com/PyLops/pylops-mpi/blob/main/tutorials_nccl/lsm_nccl.py>`_ in the repository.
+    fig3.savefig("data2.png")